hopefully fix bug where it sometimes reports the JobLock as good incorrectly
what happened was one parallel job was doing with JobLock(self.filename, ...)
to clean up the iterative locks when all the outputs exist.  Another job
reached `if self.checkoutputfiles and not self.filename.exists()`, but
didn't enter that if statement because self.filename did exist.  Then the first
job finished cleaning up and removed it, and the second job successfully opened
it.